Extract native top bar
