+++ Sources/QuillCodeApp/WorkspaceSwiftUIView.swift
@@ -1027,122 +1027,6 @@ private struct QuillCodeSearchView: View {
     }
 }
 
-private struct QuillCodeTopBarView: View {
-    var topBar: TopBarSurface
-    var commands: [WorkspaceCommandSurface]
-    @Binding var isModelPickerPresented: Bool
-    var onSetMode: (AgentMode) -> Void
-    var onSetModel: (String) -> Void
-    var onToggleModelFavorite: (String) -> Void
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    var body: some View {
-        HStack(spacing: 12) {
-            identityCluster
-                .layoutPriority(3)
-            Spacer(minLength: 10)
-            QuillCodeModelPickerView(
-                topBar: topBar,
-                isPresented: $isModelPickerPresented,
-                onSetMode: onSetMode,
-                onSetModel: onSetModel,
-                onToggleModelFavorite: onToggleModelFavorite
-            )
-                .layoutPriority(2)
-            statusIndicator
-            commandMenu
-        }
-        .padding(.horizontal, 14)
-        .padding(.vertical, 10)
-        .background(QuillCodePalette.panel)
-    }
-
-    private var identityCluster: some View {
-        HStack(spacing: 0) {
-            VStack(alignment: .leading, spacing: 3) {
-                Text(topBar.primaryTitle)
-                    .font(.headline)
-                    .lineLimit(1)
-                    .truncationMode(.tail)
-            }
-            .frame(minWidth: 0, alignment: .leading)
-            .help(topBar.subtitle)
-        }
-        .frame(minWidth: 0, alignment: .leading)
-    }
-
-    private var statusIndicator: some View {
-        HStack(spacing: 8) {
-            if showsStatusDot {
-                Circle()
-                    .fill(statusTint)
-                    .frame(width: 8, height: 8)
-                    .help(topBar.agentStatus)
-                    .accessibilityLabel("Agent status: \(topBar.agentStatus)")
-            }
-            if let runtimeIssueLabel = topBar.runtimeIssueLabel {
-                QuillCodePill(
-                    text: runtimeIssueLabel,
-                    systemImage: "exclamationmark.triangle",
-                    tint: topBar.runtimeIssueSeverity == .error ? QuillCodePalette.red : QuillCodePalette.yellow,
-                    maxWidth: 180,
-                    layoutPriority: 2
-                )
-                .help(runtimeIssueLabel)
-            }
-        }
-        .frame(minWidth: 0, alignment: .trailing)
-        .fixedSize(horizontal: true, vertical: false)
-    }
-
-    private var showsStatusDot: Bool {
-        let lowercasedStatus = topBar.agentStatus.lowercased()
-        return lowercasedStatus.contains("fail")
-            || lowercasedStatus.contains("error")
-            || lowercasedStatus.contains("run")
-            || lowercasedStatus.contains("work")
-    }
-
-    private var statusTint: Color {
-        let lowercasedStatus = topBar.agentStatus.lowercased()
-        if lowercasedStatus.contains("fail") || lowercasedStatus.contains("error") {
-            return QuillCodePalette.red
-        }
-        if lowercasedStatus.contains("run") || lowercasedStatus.contains("work") {
-            return QuillCodePalette.yellow
-        }
-        return QuillCodePalette.green
-    }
-
-    private var overflowCommands: [WorkspaceCommandSurface] {
-        TopBarOverflowCommandCatalog.commands(
-            from: commands,
-            showsComputerUseSetup: topBar.showsComputerUseSetup
-        )
-    }
-
-    private var commandMenu: some View {
-        Menu {
-            ForEach(overflowCommands) { command in
-                Button(command.title) {
-                    onCommand(command)
-                }
-                .disabled(!command.isEnabled)
-                if let shortcut = command.shortcut {
-                    Text(shortcut)
-                }
-            }
-        } label: {
-            Image(systemName: "ellipsis")
-                .frame(width: 40, height: 40)
-                .foregroundStyle(QuillCodePalette.muted)
-                .contentShape(Rectangle())
-        }
-        .buttonStyle(.borderless)
-        .help("More")
-    }
-}
-
 private struct QuillCodeSidebarView: View {
     var projects: ProjectListSurface
     var sidebar: SidebarSurface
@@ -4567,32 +4451,6 @@ private struct QuillCodeWorktreeRemoveDraft: Equatable {
             path: path.trimmingCharacters(in: .whitespacesAndNewlines),
             force: force
         )
-    }
-}
-
-private struct QuillCodePill: View {
-    var text: String
-    var systemImage: String
-    var tint: Color = QuillCodePalette.blue
-    var maxWidth: CGFloat?
-    var layoutPriority: Double = 0
-
-    var body: some View {
-        Label(text, systemImage: systemImage)
-            .font(.caption.monospacedDigit().weight(.medium))
-            .lineLimit(1)
-            .truncationMode(.tail)
-            .padding(.horizontal, 9)
-            .padding(.vertical, 6)
-            .frame(maxWidth: maxWidth, minHeight: 32)
-            .layoutPriority(layoutPriority)
-            .background(tint.opacity(0.14))
-            .foregroundStyle(tint)
-            .overlay(
-                Capsule()
-                    .stroke(tint.opacity(0.22), lineWidth: 1)
-            )
-            .clipShape(Capsule())
     }
 }
