+++ docs/DECISIONS.md
@@ -109,6 +109,7 @@
 - The sidebar's secondary tool surface is a single Tools disclosure plus a quiet Settings button. Terminal, Browser, Memories, Activity, and Command Palette are still one click away after opening Tools, but they no longer compete visually with the Codex-like primary navigation and thread list. Static HTML, SwiftUI, and Playwright share the same command IDs so this is a presentation simplification, not a second routing layer.
 - The native composer is a separate SwiftUI control instead of another private block inside `WorkspaceSwiftUIView`. Composer focus, send/stop actions, slash suggestions, and keyboard navigation now stay together, and the suggestion rows follow the shared polish contract: 40 pt minimum hit area, `0.96` press feedback, bounded command chips for long slash syntax, and quiet inline keyboard hints.
 - The native model picker is a separate SwiftUI control instead of a nested top-bar private block. The picker owns search focus, mode selection, category sections, row expansion, favorite toggles, and model selection together, with 40 pt hit targets, shared `0.96` press feedback, and middle truncation for long provider/model metadata.
+- The native top bar is a separate SwiftUI control. Workspace layout owns where the chrome sits, while `QuillCodeTopBarView` owns thread identity, model/mode access, runtime status, and the bounded utility overflow. The ellipsis uses the shared 40 pt hit target and `0.96` press feedback so the quiet Codex-like chrome still feels tactile.
 - Automations and Activity are separate surfaces. Activity is the current thread's live execution log: task plan, tools, sources, artifacts, and latest answer. Automations is for persisted or scheduled future work: thread follow-ups, workspace schedules, monitors, and later subagents. Planned workflow rows remain visible only when no configured automations exist, so the Automations pane can teach the shape of upcoming runtime work without inventing fake jobs.
 - Automation records persist as a single `~/.quillcode/automations.json` array of `QuillAutomation` values. The workspace model loads those records at bootstrap and projects them into the Automations pane, falling back to planned workflow templates only when no jobs exist. The scheduler/runner is intentionally separate from this store so heartbeat, cron, monitor, and notification execution can be added without changing the visible surface contract.
 - Automation management is command-driven. `automation-create-thread-follow-up` creates a persisted follow-up for the selected thread, `automation-create-workspace-schedule` creates a persisted schedule for the selected project, `automation-run:<id>` manually wakes runnable jobs, and row actions use `automation-pause:<id>`, `automation-resume:<id>`, and `automation-delete:<id>`. SwiftUI, static HTML, and Playwright render those command IDs from the same surface records instead of owning separate automation reducers.
